Move shuffle to util
